Add metrics ingestion scripts for GitHub API -> InfluxDB
There's various tools in here for working with PR data and turning it
into Periods for InfluxDB.  This was used to dig through the data as
part of working with it so some of the structure is based around that.

I've tried to be generic on the basis that we think we'll want to add
other data sources (eg job-server) and even pull other bits of GitHub
data into influxdb.