Switch storage over to timescaledb
This assumes the datastore will have been configured with the timescale
extension already.  Locally this is handled by the timescaledb
container, and in production we have already configured it.

The --database-url option for the CLI ensures we have that env var set.
We could always pass it through (since it's in the context) but I wasn't
sure if we actually needed it set via the CLI.

The writer class means we can ensure the appropriate table exists and
has had a hypertable created for it.

All tables have a unique constraint with the common _must_be_different
suffix so the INSERT's can act as an UPSERT (with ON CONFLICT) when it's
just the value that needs updating.