explicitly install greenlet to fix sqlalchemy install in CI
sqlalchemy==2.0.23 has a conditional install of greenlet [1] which
affects a very complex set of architectures.  This has only affected us
in CI so far, but breaks pip installs there because we don't have
greenlet installed and we're using pip in hashes mode.

The install_requires has been updated for the upcoming 2.1 release to
avoid this situation, so we can remove this change once that's released.

1: https://github.com/sqlalchemy/sqlalchemy/blob/rel_2_0_23/setup.cfg#L40