Switch PR throughput to store PRs opened&closed per day
We're trying out generating throughput based on the number of PRs closed
across a week, storing them per day, and doing the week bucketing in the
grafana queries so we can change them as needed without having to
re-ingest all the throughput data.

I've included open PRs as part of this since they may also be useful and
it is minimally more work to get those.