Replace pr-queue with open-prs command
This lets us get a count of open PRs on a given day, with a threshold
for how many days they've been open.

We want this command to only care about a 7-day threshold, but we're
keeping the other thresholds in the backfill command so we can compare
and contrast in grafana.

Note: A zero-days threshold (ie all PRs open on a day) here is the same
data that PR throughput is tracking but each command is scoped to what's
needed for that metric so we're ok with some duplication, particularly
while we're still scoping out what's needed with the stakeholders.