Add engine test fixture
We want to test some of the code against the database it's executing
against.  This fixture runs before (and then after) the full test suite,
creating a test database and dropping it when we're done.

In the case of the test db still existing it removes it so we're always
starting fresh.