Remove explicit grants assigned to the new tables
We're setting up the grafana user in timescaledb with default
permissions to select on all tables so we don't need to do this when
creating new tables.

This also frees us up to use the base timescale image in GHA as a
service, where we can't build an image in the service definition.