Get all vulnerabilities open or closed
This counts the number of vulnerability alerts that were either open or closed
for any given date and any repo. Alerts are never deleted, they are either
fixed or dismissed.

These counts are then written to a db table. For some reason, the db table
needs to have a column called "value". I'm not sure why and it's not being used
at the moment.

There is no paging on requests to the graphql api at the moment, so some repos
in the ebmdatalab org will be getting missed. In future, we will miss some
vulnerabilities too if we don't page the results.

Archived repos are being included in the responses at the moment and should be
removed.