Flatten timescaledb package
Now that writing to timescaledb is just a function there didn't seem
much need for the writer module.  This also showed up that having a
different import path for the tables was a little bit of a sore thumb.

This moves everything that one should want from the timescaledb package
into the __all__.

Notably it leaves the TIMESCALEDB_URL imports alone, in the hope that a
better method for creating an engine can be found.