Fetch all data before db table reset
The for-loop has been removed, this results in some verbose and duplicated
code that can be improved once we have a better abstraction. However, for the
moment it ensures that all data is fetched before the database table is reset.
Previously it was fetched on demand when writing the data. Unfortunately, this
means that if there's a problem with the rather flaky GitHub API then we're
left with no data, whereas this way at least we still have yesterday's data.

Some overly verbose logging statements were also removed to make running the
code locally a bit nicer.