Fetch all data before table reset
Converting the data from a generator into a list ensures that data is fetched
before the database table is reset. Previously it was fetched on demand when
writing the data. Unfortunately, this means that if there's a problem with the
rather flaky GitHub API then we're left with no data, whereas this way at least
we still have yesterday's data and some log statements.

I've also shuffled things around a bit to make some future refactorings easier.