Exclude non-Tech-owned repos
This initial deny-list includes all repos in ebmdatalab that happen to
have old PRs open right now. We can consider a more thorough list if
we get more false positives cropping up, but this seems like a cheap
way of weeding out most of the repos that we're not interested in and
which are currently affecting the charts.