Manually deploy updates to grafana
We are not able to determine when an update has been made by dependabot and then automatically deploy that change, due to `on: push:` not working with the `paths` option when there's a merge commit.

We are also unable to easily check if grafana has been updated, which means most scheduled runs of the grafana workflow will fail. It means we're unlikely to spot real errors when they happen.

Therefore, we are reverting to manual deployments for now.