Rewrite PR metrics logic
This is a complete rewrite to try to make the code simpler. I partly
refactored the old code so I could use its behaviour to drive the
design, but have deliberately deviated from how it worked where that
made the implementation simpler (for example we now record throughput
weekly rather than daily).

I've reworked some supporting code as needed rather than separating
those changes because doing so would have meant changing the existing
PR logic which was rather difficult.

In future we might want to use dataclasses for the types returned by the
queries, which would obviate the need for the FrozenDict. But that was a
larger change than I wanted to take on immediately, so the FrozenDict is
a temporary expediency to make this change possible.