Add smoke test
We should have some way of testing that the built image works before deploying
it. The previous solution of running the cli tool to output the help message is
no longer applicable and we don't want to run any of the tasks that hit an
external API, so I've added a task that only prints a log message.