Update CMD
We removed the Metrics cli, so the default entrypoint for the Docker image has
changed. This command is run by Dokku when starting the container. Dokku will
only look at whether the command is successful or not. An unsuccessful command
will result in the container being restarted, which we want to avoid.  This
command is overriden when it's invoked via a Dokku cron job.