Check in updated PR metrics queries
We don't yet have an automated way to get these out of the codebase and
into Grafana. But these queries are complicated enough that I wanted to
be able to diff them against each other and against possible changes.

The repetitive complexity of the queries suggests that we should either
do more preprocessing of the data or introduce some views, but I'd like
to defer that to next time we need to change them.