Use GitHub teams for team ownership throughout instead of hard-coded exclusion list
This has a side-effect of slightly changing the behaviour with respect
to archived repos. We don't insist that archived repos have an owning
team recorded. Now that we only display data for repos owned by
_current_ teams, any archived repos that were owned by old teams or
which were active before we had a team structure will not be included in
the data.

This doesn't affect our main use of this data, which is to understand
how the teams are operating today. It potentially affects our long-term
historic view of the data. In practice only 4% of PRs are excluded by
this change and eye-balling the resulting charts over long timescales
shows no significant differences, so we've decided that this is okay.