Scope fixture to the only module it's used in
Having a conftest isn't giving us any benefit here. And now that the
fixture is clearly scoped to the module, we reduce the pytest scope from
session to module too.

This latter makes a small improvement to devex. Previously the teardown
for the engine occurred after the last test had run, making it appear
that one non-database test was weirdly slow; now it's clear that the
slowness is due just to the database tests.