Encapsulate the GitHub API client in the query module
We never vary this, so there's no benefit to passing it around the
place.

There is a minor layering violation/duplication here because the `query`
module knows which orgs its going to be asked about. We could fix this
by 1) having a convention for mapping between org names and token env
var names and 2) having `query` dynamically creating client instances
for the orgs that it's asked about. But that seems like overkill when
this code isn't going to be reused elsewhere.