Consistent approach to importing datetime
I could be talked out of this. It's a bit more verbose and triggering
line-splits in a few places.

I originally made this change in one place in response to Dave's
feedback on a PR -- in that case it enabled the use of the natural name
for variable (date) that otherwise required mangling. Now I've decided
to be consistent.

A concrete alternative option would be to qualify datetime.date but keep
unqualified the names that are less likely to clash (timedelta,
datetime(:eyeroll:).