Distinguish content PRs
Some repos have a lot of PRs raised by people outside the tech
team. We don't necessarily want to apply the same processes to them,
so we distinguish them in the metrics.

We dynamically retrieve the membership of the tech teams from
GitHub. This means that people who leave those teams will no longer be
correctly identified, so we have a hard-coded list for them (and for
members of the tech group who contribute some code but aren't in one
of the teams). This hard-coding risks going out of date, but that's
not a big problem because we're really most interested in recent data,
for which the team lists will be correct.