Add Sentry initialisation and configuration
Based on configuration example from Sentry docs.

Timezone and schedule match current config on dokku3

Previous runs of the cron job seem to take <10mins in total.
Therefore 10 minutes runtime for each individual task should be plenty.

The first task in the list is expected to start shortly after midnight,
but the later ones might not start until nearly 10 past.

Therefore 30 mins chosen to give a bit of leeway on the checkin time.