WIP: switch to full timestamps for PR ages
Previously we only dealt in dates, so PR ages were always whole numbers
of days. However I didn't completely trust the way that interacted with
the Kaplan-Meier estimates so here we switch to more fine-grained
tracking of PR dates.

However this commit doesn't try to get the other metrics working
correctly, so they are almost certainly broken by this change.