Add bare minimum table migrations
The GitHubCodespaces table is persistent between executions of its metrics task.
We need to add columns for future features, and thus need some form of migrations.

The officially supported route for migrations in SQLAlchemy is Alembic
but for these purposes this is overkill and brings a lot of unnecessary
complexity.

This commit makes `_ensure_table`` check if columns have been added or removed
(does not support alterations to column types or nullability) and ensures
that the database schema is in sync with the table schemata defined in tables.py.

Adding `NOT NULL` columns requires that a default value specification is given.

`reset_table` requires that the drop table is committed before `_ensure_table`
is called again because of nuances of SQLAlchemy's metadata lifecycle.

Signed-off-by: Jon Massey <jon.massey@thedatalab.org>