Implement deleted flag for codespaces
If we are to ascertain which *currently active* codespaces are at risk of
deletion (due to 30d lifespan), we need to know the status of current and
past codespaces.

When deleted, a codespace's record is no long returned by the GitHub API
so we can assume that all codespaces that are not listed in the API response
are no longer in existence.