Use the same logic for calculating two PR metrics
This is very slightly slower because we give up a date range
calculation in favour of examining each day separately, but it allows
us to reuse the logic, so that seems like a reasonable trade-off.

It also introduces a slight semantic change. Previously we were
calculating an average opened _per week day_, this is now a plain
average per day. I still like the idea of just using week days, because
it means that the numbers should correspond more closely to our
experience. But I now think that the _right_ way to do that is to rework
the entire model to ignore weekends, rather than just treating this one
calculation as a special case. Doing that is quite invasive, so I'm not
sure if it will ever be worth the effort.