Dispose of database engine after testing
Prevents:

```
ResourceWarning: connection <psycopg.Connection [IDLE] (host=localhost port=5433 user=user database=test_metrics) at 0x7199eb8b9490> was deleted while still open. Please use 'with' or '.close()' to close the connection
```

Not quite sure why this warning appears for this change of dependencies,
but cleaning up the database properly fixes it.

We were using `with engine.connect() as connection` elsewhere, which
releases the connection but doesn't apparently close it:

```
When the `Connection` is closed at the end of the `with:` block,
the referenced DBAPI connection is released to the connection pool.
From the perspective of the database itself,
the connection pool will not actually "close" the connection
assuming the pool has room to store this connection for the next use.
```

https://docs.sqlalchemy.org/en/20/core/connections.html

The other alternative is using `create_engine(url, poolclass=NullPool)`
which disables connection pooling:

https://docs.sqlalchemy.org/en/20/core/pooling.html#switching-pool-implementations